feat: Add BaseComboBox and BaseTextField components
The commit adds two new QML components, BaseComboBox.qml and BaseTextField.qml, to the WeigherUI library. These components provide basic functionality for a combo box and a text field respectively. They are implemented using Qt Quick Controls 2.15 and have a fixed width and height of 200x40. The BaseComboBox.qml and BaseTextField.qml files are located in the UiComponent/BasicQmls/ComboBox and UiComponent/BasicQmls/TextField directories respectively.

The commit also includes updates to the CMakeLists.txt file to add the new QML files to the resource list.

This commit addresses the following tasks:
- Add BaseComboBox.qml
- Add BaseTextField.qml
- Update CMakeLists.txt

Note: The recent user commits and repository commits are not directly related to these code changes and are not included in the commit message.